Refactor: restructure MyMovies screen by separating components into list and upload, and enhance styling for improved user experience
