+++ src/screens/mymovies/list/index.styled.ts
